requirements.txt
streamlit==1.30.0
tensorflow
pillow==10.2.0
numpy==1.26.0

